+++ sf-proj-ios/SFPAuthorityProjections.h
@@ -38,7 +38,43 @@
  *            coordinate code
  * @return projection
  */
+-(SFPProjection *)  projectionForNumberCode: (NSNumber *) code;
+
+/**
+ * Get the projection for the code
+ *
+ * @param code
+ *            coordinate code
+ * @return projection
+ */
 -(SFPProjection *) projectionForCode: (NSString *) code;
+
+/**
+ * Check if the collection contains the projection
+ *
+ * @param projection
+ *            projection
+ * @return true if has the projection
+ */
+-(BOOL) hasProjection: (SFPProjection *) projection;
+
+/**
+ * Check if the collection contains a projection for the code
+ *
+ * @param code
+ *            coordinate code
+ * @return true if has the projection
+ */
+-(BOOL) hasProjectionNumberCode: (NSNumber *) code;
+
+/**
+ * Check if the collection contains a projection for the code
+ *
+ * @param code
+ *            coordinate code
+ * @return true if has the projection
+ */
+-(BOOL) hasProjectionCode: (NSString *) code;
 
 /**
  * Add the projection to the authority
@@ -54,19 +90,55 @@
 -(void) clear;
 
 /**
- * Clear the projection with the code
+ * Remove the projection with the code
  *
  * @param code
  *            coordinate code
  */
--(void) clearNumberCode: (NSNumber *) code;
+-(void) removeNumberCode: (NSNumber *) code;
 
 /**
- * Clear the projection with the code
+ * Remove the projection with the code
  *
  * @param code
  *            coordinate code
  */
--(void) clearCode: (NSString *) code;
+-(void) removeCode: (NSString *) code;
+
+/**
+ * Remove the projection
+ *
+ * @param projection
+ *            projection
+ */
+-(void) removeProjection: (SFPProjection *) projection;
+
+/**
+ * Get the count of authority projections
+ *
+ * @return count
+ */
+-(int) count;
+
+/**
+ * Is the collection empty?
+ *
+ * @return true if empty
+ */
+-(int) isEmpty;
+
+/**
+ * Get the projection codes
+ *
+ * @return projection codes
+ */
+-(NSArray<NSString *> *) getCodes;
+
+/**
+ * Get the projections
+ *
+ * @return projections
+ */
+-(NSArray<SFPProjection *> *) getProjections;
 
 @end