+++ sf-proj-ios/SFPAuthorityProjections.m
@@ -37,24 +37,60 @@
     return _authority;
 }
 
+-(SFPProjection *)  projectionForNumberCode: (NSNumber *) code{
+    return [self projectionForCode:[code stringValue]];
+}
+
 -(SFPProjection *) projectionForCode: (NSString *) code{
-    return [_projections objectForKey:code];
+    return [_projections objectForKey:[code uppercaseString]];
+}
+
+-(BOOL) hasProjection: (SFPProjection *) projection{
+    return [self hasProjectionCode:[projection code]];
+}
+
+-(BOOL) hasProjectionNumberCode: (NSNumber *) code{
+    return [self hasProjectionCode:[code stringValue]];
+}
+
+-(BOOL) hasProjectionCode: (NSString *) code{
+    return [self projectionForCode:code];
 }
 
 -(void) addProjection: (SFPProjection *) projection{
-    [_projections setObject:projection forKey:projection.code];
+    [_projections setObject:projection forKey:[projection.code uppercaseString]];
 }
 
 -(void) clear{
     [_projections removeAllObjects];
 }
 
--(void) clearNumberCode: (NSNumber *) code{
-    [self clearCode:[code stringValue]];
+-(void) removeNumberCode: (NSNumber *) code{
+    [self removeCode:[code stringValue]];
 }
 
--(void) clearCode: (NSString *) code{
-    [_projections removeObjectForKey:code];
+-(void) removeCode: (NSString *) code{
+    [_projections removeObjectForKey:[code uppercaseString]];
+}
+
+-(void) removeProjection: (SFPProjection *) projection{
+    [self removeCode:[projection code]];
+}
+
+-(int) count{
+    return (int) self.projections.count;
+}
+
+-(int) isEmpty{
+    return [self count] == 0;
+}
+
+-(NSArray<NSString *> *) getCodes{
+    return [self.projections allKeys];
+}
+
+-(NSArray<SFPProjection *> *) getProjections{
+    return [self.projections allValues];
 }
 
 @end