+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 2.0.3 (TBD)
 
 * sf-ios 2.0.3
+* Added OGC projection authority
+* Projections collection of cached authority projections
+* Create uppercase projection authorities
 
 ## [2.0.2](https://github.com/ngageoint/simple-features-proj-ios/releases/tag/2.0.2) (04-03-2019)